+++ examples/with-solana/README.md
@@ -166,7 +166,7 @@ Token balance for warchest: 0.0001
 
 Enjoy!
 
-### 6/ Running the "Allow SPL token transfer policy" example
+### 6/ Running the "Create SPL token transfer policy" example
 
 This example shows the flow of using our policy engine's support for SPL token transfers! Specifically we show how to allow a user (non root user) to send a Solana SPL token transfer by calculating the "associated token address" of the receiving account, given the mint address of the token that we are transferring.
 
@@ -177,14 +177,14 @@ To best see how policies can be used effectively, we will go through example in 
 3. Create Token Policy - this will create a policy that allows Solana transactions initiated by the correct user to be signed if they contain a single instruction which is an SPL token transfer to the correct token account
 4. Attempt Transfer (again) - this will succeed, now that the appropriate policy has been created!
 
-NOTE: An very key piece of understanding to glean from this example is the use of "associated token addresses (or "associated token account address") in the creation of SPL related policies. For further context on ATA's look here: https://spl.solana.com/associated-token-account
+NOTE: A very key piece of understanding to glean from this example is the use of "associated token addresses (or "associated token account address") in the creation of SPL related policies. For further context on ATA's look here: https://spl.solana.com/associated-token-account
 
 #### Step 1. Setup
 
 First we will run the setup command
 
 ```bash
-$ pnpm allow-token-transfer setup
+$ pnpm token-transfer-policy setup
 ```
 
 This will create a Solana wallet address that you'll have to fund with devnet SOL via a faucet or cli. A commonly used SOL devnet faucet is available here --> https://faucet.solana.com/
@@ -228,7 +228,7 @@ Non root user created with user id: 13c4609b-8818-40be-aeac-c8b63da0de4a
 We will now attempt to transfer these SPL tokens that we've created
 
 ```bash
-$ pnpm allow-token-transfer attempt_transfer
+$ pnpm token-transfer-policy attempt_transfer
 ```
 
 This will prompt you first to enter in the Solana wallet address that is originating the transfer (use the address labeled "Turnkey Solana wallet address" output at the end of the setup step)
@@ -261,7 +261,7 @@ details: [
 This is the most important step! In this step we are taking our receiving address (the Turnkey Warchest), calculating it's "associated token address", given the Mint account address for the token we've created, and creating a policy that allows the correct non root user to sign Solana transactions if and only if they contain a single instruction that is an SPL token transfer to this associated token address.
 
 ```bash
-$ pnpm allow-token-transfer create_token_policy
+$ pnpm token-transfer-policy create_token_policy
 ```
 
 This will prompt you first to enter in the user ID of the non root user who's credentials are being used to sign the transfer (use the value labeled "Non root user created with user id" output at the end of the setup step)
