+++ examples/with-solana/package.json
@@ -6,7 +6,7 @@
     "start": "pnpm tsx src/index.ts",
     "advanced": "pnpm tsx src/advanced.ts",
     "token-transfer": "pnpm tsx src/tokenTransfer.ts",
-    "allow-token-transfer": "pnpm tsx src/allowTokenTransferPolicy.ts",
+    "token-transfer-policy": "pnpm tsx src/tokenTransferPolicy.ts",
     "faucet": "pnpm tsx src/faucet.ts",
     "jupiter-swap": "pnpm tsx src/jupiterSwap.ts",
     "with-fee-payer": "pnpm tsx src/withFeePayer.ts",
