+++ examples/with-solana/src/utils/index.ts
@@ -3,12 +3,11 @@ import * as solanaNetwork from "./solanaNetwork";
 import { signMessage } from "./signMessage";
 import { createTransfer } from "./createSolanaTransfer";
 import { print } from "./print";
-import { isKeyOfObject } from "./isKeyOfObject";
 import { createMint } from "./createMint";
 import { createToken } from "./createToken";
 import { createTokenAccount } from "./createTokenAccount";
 import {
-  createTokenTransfer,
+  createTokenTransferAddSignature,
   createTokenTransferSignTransaction,
 } from "./createTokenTransfer";
 import { transactionSenderAndConfirmationWaiter } from "./retrySender";
@@ -21,14 +20,34 @@ export {
   createNewSolanaWallet,
   createToken,
   createTokenAccount,
-  createTokenTransfer,
+  createTokenTransferAddSignature,
   createTokenTransferSignTransaction,
   createTransfer,
   handleActivityError,
   print,
-  isKeyOfObject,
   signMessage,
   transactionSenderAndConfirmationWaiter,
   solanaNetwork,
   TURNKEY_WAR_CHEST,
 };
+
+export function refineNonNull<T>(
+  input: T | null | undefined,
+  errorMessage?: string
+): T {
+  if (input == null) {
+    throw new Error(errorMessage ?? `Unexpected ${JSON.stringify(input)}`);
+  }
+
+  return input;
+}
+
+// isKeyOfObject checks if a key exists within an object
+export function isKeyOfObject<T>(
+  key: string | number | symbol | undefined,
+  obj: any
+): key is keyof T {
+  if (!key) return false;
+
+  return key in obj;
+}
